fully integrating b+ string tree

+++ Project_2/src/bpt_string.cpp
@@ -1284,17 +1284,54 @@ string_node * Bpt_string::delete_entry( string_node * n, string key, void * poin
 
 /* Master deletion function.
  */
-string_node * Bpt_string::delete_node( string key ) {
-
-	string_node * key_leaf;
-	record * key_record;
+string_node * Bpt_string::delete_node( string key, int value ) {
+	
+    string_node * key_leaf;
+	record * key_record, * temp, *temp2;
+    bool delete_from_tree = false;
+    int i;
 
 	key_record = find(key);
 	key_leaf = find_leaf(key);
-	if (key_record != NULL && key_leaf != NULL) {
-		root = delete_entry(key_leaf, key, key_record);
-		free(key_record);
-	}
+    temp = key_record;
+    /* only delete the record we want */
+    /* case 1: only 1 record */
+    if (temp->next == NULL && temp->value == value) {
+        delete_from_tree = true;
+    }
+    /* case 2: mult records, first one is one we want */
+    else if (temp->next != NULL && temp->value == value) {
+        delete_from_tree = false;
+        //for (i = 0; i < key_leaf->num_keys; i++)
+        //    if (key_leaf->keys[i] == key) break;
+        i = 0;
+        for (iter = key_leaf->keys.begin(); iter != key_leaf->keys.end(); ++iter, i++)
+            if (*iter == key) break;
+        
+        key_leaf->pointers[i] = temp->next;
+        free(temp);
+
+    }
+    /* case 3: mult records, first one is not what we want */
+    else {
+        delete_from_tree = false;
+        while (temp->next) {
+            temp2 = temp;
+            temp = temp->next;
+            if (temp->value == value) {
+                temp2->next = temp->next;
+                free(temp);
+                break;
+            }
+        }
+    }
+
+    if (delete_from_tree) {
+        if (key_record != NULL && key_leaf != NULL) {
+            root = delete_entry(key_leaf, key, key_record);
+            free(key_record);
+        }
+    }
 	return root;
 }
